// src/app/pages/auth/types/Types.ts
export type SignUpFormInputs = {
    email: string;
    password: string;
    name: string;
    lastname: string;
  };

 export type SignInFormInputs = {
    email: string;
    password: string;
  };
  
  export type SignUpDemo = {
    email: string;
    password: string;
    displayName: string;
  };